+++ src/compose/compose-scalar.ts
@@ -1,7 +1,7 @@
-import { isScalar } from '../ast/Node.js'
-import { Scalar } from '../ast/Scalar.js'
 import type { Document } from '../doc/Document.js'
 import type { Schema } from '../doc/Schema.js'
+import { isScalar } from '../nodes/Node.js'
+import { Scalar } from '../nodes/Scalar.js'
 import type { BlockScalar, FlowScalar } from '../parse/tokens.js'
 import type { ScalarTag } from '../tags/types.js'
 import { resolveBlockScalar } from './resolve-block-scalar.js'
