+++ src/compose/resolve-block-map.ts
@@ -1,7 +1,7 @@
-import { Pair } from '../ast/Pair.js'
-import { YAMLMap } from '../ast/YAMLMap.js'
 import { Type } from '../constants.js'
 import type { Document } from '../doc/Document.js'
+import { Pair } from '../nodes/Pair.js'
+import { YAMLMap } from '../nodes/YAMLMap.js'
 import type { BlockMap } from '../parse/tokens.js'
 import type { ComposeNode } from './compose-node.js'
 import { resolveMergePair } from './resolve-merge-pair.js'
