+++ src/compose/resolve-flow-collection.ts
@@ -1,9 +1,9 @@
-import { isNode, isPair, ParsedNode } from '../ast/Node.js'
-import { Pair } from '../ast/Pair.js'
-import { YAMLMap } from '../ast/YAMLMap.js'
-import { YAMLSeq } from '../ast/YAMLSeq.js'
 import { Type } from '../constants.js'
 import type { Document } from '../doc/Document.js'
+import { isNode, isPair, ParsedNode } from '../nodes/Node.js'
+import { Pair } from '../nodes/Pair.js'
+import { YAMLMap } from '../nodes/YAMLMap.js'
+import { YAMLSeq } from '../nodes/YAMLSeq.js'
 import type { FlowCollection, SourceToken, Token } from '../parse/tokens.js'
 import type { ComposeNode } from './compose-node.js'
 import { resolveEnd } from './resolve-end.js'
