+++ src/doc/createNode.ts
@@ -1,8 +1,8 @@
-import type { Alias } from '../ast/Alias.js'
-import { isNode, isPair, Node } from '../ast/Node.js'
-import { Scalar } from '../ast/Scalar.js'
-import type { YAMLMap } from '../ast/YAMLMap.js'
 import { defaultTagPrefix } from '../constants.js'
+import type { Alias } from '../nodes/Alias.js'
+import { isNode, isPair, Node } from '../nodes/Node.js'
+import { Scalar } from '../nodes/Scalar.js'
+import type { YAMLMap } from '../nodes/YAMLMap.js'
 import type { TagObj } from '../tags/types.js'
 import type { Replacer } from './Document.js'
 import type { Schema } from './Schema.js'
