+++ src/nodes/Scalar.ts
