+++ src/stringify/stringifyString.ts
@@ -1,5 +1,5 @@
-import type { Scalar } from '../ast/Scalar.js'
 import { Type } from '../constants.js'
+import type { Scalar } from '../nodes/Scalar.js'
 import { strOptions } from '../tags/options.js'
 import { addCommentBefore } from './addComment.js'
 import {
