+++ src/tags/failsafe/map.ts
@@ -1,6 +1,6 @@
-import { isMap } from '../../ast/Node.js'
-import { createPair } from '../../ast/Pair.js'
-import { YAMLMap } from '../../ast/YAMLMap.js'
+import { isMap } from '../../nodes/Node.js'
+import { createPair } from '../../nodes/Pair.js'
+import { YAMLMap } from '../../nodes/YAMLMap.js'
 import type { CreateNodeContext } from '../../doc/createNode.js'
 import type { Schema } from '../../doc/Schema.js'
 import type { CollectionTag } from '../types.js'
