+++ src/tags/types.ts
@@ -1,9 +1,9 @@
-import type { Node } from '../ast/Node.js'
-import type { Scalar } from '../ast/Scalar.js'
-import type { YAMLMap } from '../ast/YAMLMap.js'
-import type { YAMLSeq } from '../ast/YAMLSeq.js'
 import type { CreateNodeContext } from '../doc/createNode.js'
 import type { Schema } from '../doc/Schema.js'
+import type { Node } from '../nodes/Node.js'
+import type { Scalar } from '../nodes/Scalar.js'
+import type { YAMLMap } from '../nodes/YAMLMap.js'
+import type { YAMLSeq } from '../nodes/YAMLSeq.js'
 import type { StringifyContext } from '../stringify/stringify.js'
 
 export type SchemaId = 'core' | 'failsafe' | 'json' | 'yaml11'
