+++ src/tags/yaml-1.1/binary.ts
@@ -1,7 +1,5 @@
-/* global atob, btoa, Buffer */
-
-import type { Scalar } from '../../ast/Scalar.js'
 import { Type } from '../../constants.js'
+import type { Scalar } from '../../nodes/Scalar.js'
 import { stringifyString } from '../../stringify/stringifyString.js'
 import { binaryOptions as options } from '../options.js'
 import type { ScalarTag } from '../types.js'
