+++ src/tags/yaml-1.1/set.ts
@@ -1,9 +1,9 @@
-import { isMap, isPair, isScalar } from '../../ast/Node.js'
-import { createPair, Pair } from '../../ast/Pair.js'
-import { Scalar } from '../../ast/Scalar.js'
-import { ToJSContext } from '../../ast/toJS.js'
-import { YAMLMap, findPair } from '../../ast/YAMLMap.js'
 import type { Schema } from '../../doc/Schema.js'
+import { isMap, isPair, isScalar } from '../../nodes/Node.js'
+import { createPair, Pair } from '../../nodes/Pair.js'
+import { Scalar } from '../../nodes/Scalar.js'
+import { ToJSContext } from '../../nodes/toJS.js'
+import { YAMLMap, findPair } from '../../nodes/YAMLMap.js'
 import type { StringifyContext } from '../../stringify/stringify.js'
 import type { CollectionTag } from '../types.js'
