+++ src/util.ts
@@ -1,5 +1,5 @@
-export { findPair } from './ast/YAMLMap.js'
-export { toJS, ToJSContext } from './ast/toJS.js'
+export { findPair } from './nodes/YAMLMap.js'
+export { toJS, ToJSContext } from './nodes/toJS.js'
 export { stringifyNumber } from './stringify/stringifyNumber.js'
 export { stringifyString } from './stringify/stringifyString.js'
 export { Type } from './constants.js'
